requirements.txt
boto3>=1.34.0
prompt_toolkit>=3.0.43
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0